refactor: model 패키지
