+++ RemindAlarm/app/src/main/java/com/impactrudia/remindalarm/presentation/model/AlarmModel.kt
@@ -1,4 +1,4 @@
-package com.impactrudia.remindalarm.presentation.ui.domain
+package com.impactrudia.remindalarm.presentation.model
 
 class AlarmModel(val id: Int, val memo: String) : MyItem() {
 }
