+++ RemindAlarm/app/src/main/java/com/impactrudia/remindalarm/presentation/model/RingtoneModel.kt
@@ -1,4 +1,4 @@
-package com.impactrudia.remindalarm.presentation.ui.domain
+package com.impactrudia.remindalarm.presentation.model
 
 class RingtoneModel(val id: Int, val title: String) : MyItem() {
 }
