+++ RemindAlarm/app/src/main/java/com/impactrudia/remindalarm/presentation/model/enums/AdapterType.kt
@@ -1,0 +1,5 @@
+package com.impactrudia.remindalarm.presentation.model.enums
+
+enum class AdapterType {
+    ALARM, RINGTONE
+}
