--- RemindAlarm/app/src/main/java/com/impactrudia/remindalarm/presentation/ui/domain/enum/AdapterType.kt
@@ -1,5 +1,0 @@
-package com.impactrudia.remindalarm.presentation.ui.domain.enum
-
-enum class AdapterType {
-    ALARM, RINGTONE
-}
