+++ RemindAlarm/app/src/main/java/com/impactrudia/remindalarm/presentation/ui/main/RingtoneSelectionFragment.kt
@@ -1,26 +1,23 @@
 package com.impactrudia.remindalarm.presentation.ui.main
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
+import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.DividerItemDecoration
 import com.impactrudia.remindalarm.R
 import com.impactrudia.remindalarm.databinding.FragmentRingtoneSelectionBinding
+import com.impactrudia.remindalarm.presentation.model.RingtoneModel
+import com.impactrudia.remindalarm.presentation.model.enums.AdapterType
 import com.impactrudia.remindalarm.presentation.ui.base.RecyclerViewAdapter
-import com.impactrudia.remindalarm.presentation.ui.domain.RingtoneModel
-import com.impactrudia.remindalarm.presentation.ui.domain.enum.AdapterType
 
 class RingtoneSelectionFragment : Fragment() {
 
     private lateinit var binding: FragmentRingtoneSelectionBinding
-
     private val ringtoneAdapter by lazy {
         RecyclerViewAdapter(AdapterType.RINGTONE) {
             if (it is RingtoneModel) {
-                Toast.makeText(requireActivity(), "title::${it?.title}", Toast.LENGTH_SHORT).show()
             }
         }
     }
